renamed files in /nav Navbar.svelte is root, everything else is ancillary and forced below
