+++ src/kong_svelte/src/lib/components/nav/NavbarOption.svelte
