+++ src/kong_svelte/src/lib/components/nav/NavbarPanel.svelte
